requirements.txt
python-dotenv==1.0.1
pandas==2.2.3
PyMuPDF==1.25.3
sqlalchemy==2.0.38
google-generativeai==0.8.3
reportlab==4.2.5
